combine the serde for message and payload
