+++ p2p/src/main/kotlin/maru/p2p/messages/BeaconBlocksByRangeResponseMessageSerDe.kt
@@ -8,17 +8,44 @@
  */
 package maru.p2p.messages
 
+import maru.core.SealedBeaconBlock
 import maru.p2p.Message
 import maru.p2p.RpcMessageType
 import maru.p2p.Version
 import maru.serialization.SerDe
+import maru.serialization.rlp.SealedBeaconBlockSerDe
+import org.apache.tuweni.bytes.Bytes
+import org.hyperledger.besu.ethereum.rlp.RLP
+import org.hyperledger.besu.ethereum.rlp.RLPInput
+import org.hyperledger.besu.ethereum.rlp.RLPOutput
 
 class BeaconBlocksByRangeResponseMessageSerDe(
-  private val responseSerDe: SerDe<BeaconBlocksByRangeResponse>,
+  private val sealedBeaconBlockSerDe: SealedBeaconBlockSerDe,
 ) : SerDe<Message<BeaconBlocksByRangeResponse, RpcMessageType>> {
   override fun serialize(value: Message<BeaconBlocksByRangeResponse, RpcMessageType>): ByteArray =
-    responseSerDe.serialize(value.payload)
+    RLP.encode { rlpOutput ->
+      writeTo(value.payload, rlpOutput)
+    }.toArray()
 
   override fun deserialize(bytes: ByteArray): Message<BeaconBlocksByRangeResponse, RpcMessageType> =
-    Message(RpcMessageType.BEACON_BLOCKS_BY_RANGE, Version.V1, responseSerDe.deserialize(bytes))
+    Message(RpcMessageType.BEACON_BLOCKS_BY_RANGE, Version.V1, readFrom(RLP.input(Bytes.wrap(bytes))))
+
+  private fun writeTo(
+    value: BeaconBlocksByRangeResponse,
+    rlpOutput: RLPOutput,
+  ) {
+    rlpOutput.startList()
+    rlpOutput.writeList(value.blocks) { block, output ->
+      sealedBeaconBlockSerDe.writeTo(block, output)
+    }
+    rlpOutput.endList()
+  }
+
+  private fun readFrom(rlpInput: RLPInput): BeaconBlocksByRangeResponse {
+    rlpInput.enterList()
+    val blocks = rlpInput.readList { sealedBeaconBlockSerDe.readFrom(it) }
+    rlpInput.leaveList()
+
+    return BeaconBlocksByRangeResponse(blocks = blocks)
+  }
 }
